requirements.txt
pyyaml
flask
requests
paho-mqtt
